Added the beginnings of Android wear support.
Wear now shows notifications and vibrates when the alarm goes off.

It actually goes off after the alarm starts ringing so that could be a problem.

The bitmap shown on the watch is still way too big but I cannot get it to decrease.

My attempts to add actions such as snooze, dismiss, and silence have not been succesful yet.